ORM is working now but still WIP for PUT and DELETE requests

+++ src/main/java/com/virtech/spacez/repositories/SatelliteRepository.java
@@ -4,8 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.virtech.spacez.entities.Satellite;
 
-// This will be AUTO IMPLEMENTED by Spring into a Bean called satelliteRepository
 
-public interface SatelliteRepository extends JpaRepository<Satellite, Long> {
+public interface SatelliteRepository extends JpaRepository<Satellite, Integer> {
     
 }
